money_tracker_backend/requirements.txt
asgiref
attrs
Django
django-cors-headers
djangorestframework
gunicorn
Jinja2
Pillow
python-dateutil
requests
setuptools